refactor: introducing sessionStore as central store providing all useful session data
